pyproject.toml
[build-system]
requires = ["setuptools>=46.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trivialscan-lambda"
version = "1.0.0"
description = "Just an API"
authors = [
    {name = "Christopher Langton", email = "chris@trivialsec.com"},
]
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "validators==0.18.2",
    "anyio==3.6.1",
    "fastapi==0.79.0",
    "idna==3.3",
    "pydantic==1.9.2",
    "sniffio==1.2.0",
    "starlette==0.19.1",
    "typing-extensions==4.3.0",
    "mangum==0.15.1",
    "retry==0.9.2",
    "python-multipart==0.0.5",
    "requests==2.28.1",
    "sendgrid==6.9.7",
    "pydantic[email]==1.9.2",
    "user-agents==2.2.0",
    "geocoder==1.38.1",
    "stripe==4.2.0",
    "cachier==1.5.4",
    "tldextract==3.4.0",
    "pyyaml==6.0",
]

[tool.coverage.run]
source = ["src"]

[tool.pytest.ini_options]
addopts = "--cov --cov-report html --cov-report term-missing --cov-fail-under 95"

[tool.pylint]
max-line-length = 88
disable = [
    "C0114", # (missing-module-docstring)
    "C0115", # (missing-class-docstring)
    "C0116", # (missing-function-docstring)
    "R0903", # (too-few-public-methods)
    "R0913", # (too-many-arguments)
    "W0105", # (pointless-string-statement)
]

[tool.pylint.'MESSAGES CONTROL']
extension-pkg-whitelist = "pydantic"
